refactored PersonCard to split the elements into their own file, refactored getPersonImage to be a utility function
