+++ features/persons/components/PersonCard/index.ts
@@ -1,0 +1,1 @@
+export * from './PersonCard';
