+++ features/persons/components/PersonCard/types.ts
@@ -1,0 +1,8 @@
+export type Person = {
+  id: string;
+  name: string;
+  age: number;
+  gender: string;
+  address: string;
+  imageUrl?: string;
+};
